backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Qdrant client for vector database
qdrant-client>=1.7.0

# CLIP model dependencies
torch>=2.0.0
torchvision>=0.15.0
openai-clip

# Image processing
pillow>=10.0.0
numpy>=1.24.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.12.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.25.0

# Logging and monitoring
loguru>=0.7.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Optional: For production
gunicorn>=21.2.0
